+++ st2client/st2client/models/__init__.py
@@ -16,6 +16,7 @@
 from st2client.models.core import *         # noqa
 from st2client.models.auth import *       # noqa
 from st2client.models.action import *       # noqa
+from st2client.models.action_alias import *  # noqa
 from st2client.models.keyvalue import *    # noqa
 from st2client.models.policy import *       # noqa
 from st2client.models.reactor import *      # noqa