+++ st2tests/st2tests/config.py
@@ -52,6 +52,7 @@ def _register_config_opts():
     _register_auth_opts()
     _register_action_sensor_opts()
     _register_mistral_opts()
+    _register_scheduler_opts()
 
 
 def _override_db_opts():
@@ -150,6 +151,16 @@ def _register_mistral_opts():
         cfg.IntOpt('retry_wait', default=1)
     ]
     _register_opts(mistral_opts, group='mistral')
+
+
+def _register_scheduler_opts():
+    scheduler_opts = [
+        cfg.IntOpt('delayed_execution_recovery', default=600,
+                   help='The time in seconds to wait before recovering delayed action executions.'),
+        cfg.IntOpt('rescheduling_interval', default=300,
+                   help='The frequency for rescheduling action executions.')
+    ]
+    _register_opts(scheduler_opts, group='scheduler')
 
 
 def _register_opts(opts, group=None):